+++ src/main/java/memethespire/relicmemes/RelicModification.java
@@ -44,37 +44,18 @@ public class RelicModification {
         return relicStrings != null && relicStrings.NAME.equalsIgnoreCase(relicName);
     }
 
-    // TODO: Consider the necessity of modifying name and description
-    //  properties.
-    //
-    // The name and description properties seem to only be used when the
-    // player right clicks (views) the relic, as the PowerTips are generated
-    // separately. Modifying these properties would allow the player to see
-    // the meme when they "take a close up look" at the relic as well,
-    // however, it may not be desirable for them to see the memes when they
-    // actually want to check what the relic does.
     public void modify(AbstractRelic relic) {
-        if (modifiedName != null) {
-            ReflectionUtils.setFinal(relic, AbstractRelic.class, "name", modifiedName);
-        }
-        if (modifiedDescription != null) {
-            relic.description = modifiedDescription;
-        }
-
         // Update the power tip.
         relic.tips.clear();
-        relic.tips.add(new PowerTip(relic.name, relic.description));
+        relic.tips.add(new PowerTip(modifiedName, modifiedDescription));
 
         ReflectionUtils.invokePrivate(relic, AbstractRelic.class, "initializeTips");
     }
 
     public static void restore(AbstractRelic relic) {
-        ReflectionUtils.setFinal(relic, AbstractRelic.class, "name", getRelicStrings(relic).NAME);
-        relic.description = relic.getUpdatedDescription();
-
         // Update the power tip.
         relic.tips.clear();
-        relic.tips.add(new PowerTip(relic.name, relic.description));
+        relic.tips.add(new PowerTip(getRelicStrings(relic).NAME, relic.getUpdatedDescription()));
 
         ReflectionUtils.invokePrivate(relic, AbstractRelic.class, "initializeTips");
     }