+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+#### Changes
+
+* `RelicModification` no longer changes the name and description of the relic (it only changes the relic's power tips now).
+
 #### Content
 
 * Added 2 more silent card memes.